Flow Rates App
This is my very first program that I have written besides "hello world" and other 'getting started' style programs. My family's business is building, testing, and maintaining fuel stations. for a few sites we do monthly flow rates for their gasoline. Flow Rates must be calculated for each grade of fuel at each nozzle. therefore, a site with 24 nozzles that has 87, 89, and 91 has 72 times that fuel must be pumped, timed, and the calculations done to determine gallons per minute. My hope is that I can turn this simple python program into an app that has only a few buttons and options. Ideally a way to change how many gallons you want to pump in case people like to pump different amounts, a build in timer that automatically places the time into the calculation to give you the flow rate, and a way to restart the process. I am new to programming so it is a learning experience.